+++ search.go
@@ -744,6 +744,8 @@ func cmdSearch(db *database, inv *invocation, stdout, stderr io.Writer) int {
 		tsvOut:     o.tsvOut,
 		color:      searchColorEnabled(stdout, o.jsonOut || o.tsvOut),
 		maxColumns: o.maxColumns,
+		headSet:    o.headSet,
+		headLines:  o.headLines,
 	}
 
 	var firstErr error